+++ font/types.h
@@ -18,8 +18,8 @@ struct RasterizedGlyph {
 
 // TODO: Consider making a global "geometry" namespace and moving Point there.
 struct Point {
-    int x;
-    int y;
+    float x;
+    float y;
 };
 
 struct ShapedGlyph {
@@ -36,7 +36,7 @@ struct ShapedRun {
 
 struct LineLayout {
     size_t layout_font_id;
-    int width;
+    float width;
     size_t length;
     std::vector<ShapedRun> runs;
     int ascent;
