+++ gui/text_system/caret.h
@@ -8,8 +8,8 @@ namespace gui {
 struct Caret {
     size_t index;
 
-    size_t columnAtX(const font::LineLayout& layout, int x, bool exclude_end = false) const;
-    int xAtColumn(const font::LineLayout& layout, size_t col, bool exclude_end = false) const;
+    size_t columnAtX(const font::LineLayout& layout, float x, bool exclude_end = false) const;
+    float xAtColumn(const font::LineLayout& layout, size_t col, bool exclude_end = false) const;
 
     size_t moveToPrevGlyph(const font::LineLayout& layout, size_t col);
     size_t moveToNextGlyph(const font::LineLayout& layout, size_t col);
