+++ gui/text_system/line_layout_cache.h
@@ -9,13 +9,13 @@ namespace gui {
 class LineLayoutCache {
 public:
     const font::LineLayout& getLineLayout(std::string_view str8);
-    int maxWidth() const;
+    float maxWidth() const;
 
 private:
     std::unordered_map<XXH64_hash_t, font::LineLayout> cache;
 
     // TODO: Use a data structure (priority queue) for efficient updating.
-    int max_width = 0;
+    float max_width = 0;
 };
 
 }
